경매등록시 카테고리가 중간테이블에 저장되도록 수정

+++ src/main/java/com/mutsa/springboot_auction/domain/auction/dto/AuctionRequest.java
@@ -41,7 +41,7 @@ public class AuctionRequest {
 
     private List<String> imageUrls;
 
-    private List<Integer> categoryIds;
+    private List<Long> categoryIds;
 
     @NotNull(message = "입찰가 숨김 여부를 입력해주세요.")
     private Boolean hideBidPrice;
